// generated/src/app/api-models/cap-elgcnt-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapElgcntWork {
  seqCcalcId: number;
  capEntityCode: string;
  capEntityId: string;
  seqCapProvId: number;
  seqCapVendId: number;
  eligCntCurr: number;
  eligCnt1Back: number;
  eligCnt2Back: number;
  eligCnt3Back: number;
  eligCnt4Back: number;
  eligCnt5Back: number;
  eligCnt6Back: number;
  eligCnt7Back: number;
  eligCnt8Back: number;
  eligCnt9Back: number;
  eligCnt10Back: number;
  eligCnt11Back: number;
  maxCnt: number;
  seqCapElgcntWork: number;
  seqCapVendAddress: number;
}